src/requirements.txt
aiofiles==23.2.1
aiogram==3.12.0
aiohappyeyeballs==2.3.5
aiohttp==3.10.2
aiosignal==1.3.1
aiosqlite==0.20.0
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
bitarray==2.9.2
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
crc16==0.1.1
crc32c==2.6
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.111.1
fastapi-cli==0.0.4
frozenlist==1.4.1
graphql-core==3.2.3
graphql-relay==3.2.0
graphql_query==1.0.3
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
magic-filter==1.0.12
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyNaCl==1.5.0
python-dotenv==1.0.1
python-multipart==0.0.9
pytonlib==0.0.63
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
setuptools==72.1.0
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.37.2
ton==0.26
tonsdk==1.0.15
TonTools==2.1.2
tvm-valuetypes==0.0.12
typer==0.12.3
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.3
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
yarl==1.9.4
